requirements_test.txt
pytest-homeassistant-custom-component==0.4.0
black==21.9b0
flake8==3.9.2
isort==5.9.3
mypy==0.910
pylint==2.11.1
pylint-strict-informational==0.1
pre-commit==2.15.0
